# server/openapi.yaml
openapi: 3.1.0
info:
  title: evcc
  description: Solar charging. Super simple.
  version: "0.2.0"
  contact:
    url: https://github.com/evcc-io/evcc
servers:
  - url: https://demo.evcc.io/api
...
paths:
  /auth/login:
    post:
      operationId: login
      summary: Login
      description: "Administrator login. Returns authorization cookie required for all protected endpoints."
      tags:
        - auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                password:
                  $ref: "#/components/schemas/Password"
      responses:
        "200":
          description: Success
          headers:
            Set-Cookie:
              description: The authorization cookie
              schema:
                type: string
                example: >-
                  auth=ey...qY; Path=/; Expires=Fri, 04 Apr 2025 14:35:37 GMT;
                  HttpOnly; SameSite=Strict
        "401":
          description: Invalid password provided
...
  /auth/password:
    put:
      operationId: changePassword
      summary: Change admin password
      description: "Changes the admin password."
      tags:
        - auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePassword"
      responses:
        "202":
          description: Success
        "400":
          description: Invalid password provided
...
  /auth/apikey:
    get:
      operationId: getApiKeyStatus
      summary: API key status
      description: |
        Reports whether an API key has been generated. The key itself is
        only returned once at creation time (POST), never on subsequent
        reads.
      tags:
        - auth
      security:
        - cookieAuth: []
        - bearerAuth: []
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  configured:
                    type: boolean
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: regenerateApiKey
      summary: Generate or rotate the API key
      description: |
        Generates a fresh API key, replacing any existing one. The
        returned key is shown only once; store it immediately.

        Even when the request is authenticated via API key (Bearer), the
        admin password must be supplied in the request body to prevent a
        leaked key from rotating itself. The password check is skipped
        when the server is started with `--disable-auth`.
      tags:
        - auth
      security:
        - cookieAuth: []
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                password:
                  $ref: "#/components/schemas/Password"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  key:
                    type: string
                    description: The new API key (cleartext, shown once).
                    example: evcc_aB3xYz7k0Pq2sN5mWtR4
        "401":
          description: Invalid admin password
        "403":
          description: Forbidden in demo mode
  /batterydischargecontrol/{enable}:
    post:
      operationId: setBatteryDischargeControl
      summary: Control battery-discharge
      description: "Prevent home battery discharge during vehicle fast charging."
      externalDocs:
        url: https://docs.evcc.io/en/features/battery
      tags:
        - battery
      parameters:
        - $ref: "#/components/parameters/enable"
      responses:
        "200":
          $ref: "#/components/responses/BooleanResult"
...
  /batterygridchargelimit/{cost}:
    post:
      operationId: setBatteryGridChargeLimit
      summary: Set battery grid charge limit
      description: "Charge home battery from grid when price or emissions are below the threshold. Uses price if a dynamic tariff exists. Uses emissions if a CO₂-tariff is configured. Ignored otherwise."
      externalDocs:
        url: https://docs.evcc.io/en/features/battery
      tags:
        - battery
      parameters:
        - $ref: "#/components/parameters/costLimit"
      responses:
        "200":
          $ref: "#/components/responses/NumberResult"
...
  /batterygriddischargelimit/{cost}:
    post:
      operationId: setBatteryGridDischargeLimit
      summary: Set battery grid discharge limit
      description: "Discharge home battery to grid when the feed-in price is at or above the threshold (experimental). Requires a feed-in tariff and the battery grid discharge setting. Ignored otherwise."
      externalDocs:
        url: https://docs.evcc.io/en/features/battery
      tags:
        - battery
      parameters:
        - $ref: "#/components/parameters/costLimit"
      responses:
        "200":
          $ref: "#/components/responses/NumberResult"
  /batterypvcontrol:
    post:
      operationId: resolveBatteryPVControl
      summary: Resolve a battery PV control conflict
      description: Persist an explicit decision for the displayed journal revision. Recovery is reconciled during the next control cycle.
      tags:
        - battery
      security:
        - cookieAuth: []
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [battery, action, revision]
              properties:
                battery:
                  type: string
                action:
                  type: string
                  enum: [disable, recover]
                revision:
                  type: string
      responses:
        "200":
          description: Decision persisted
        "400":
          description: Invalid decision
        "401":
          description: Authentication required
        "409":
          description: Stale revision, unavailable journal or recovery refused
...
  /buffersoc/{soc}:
    post:
      operationId: setBufferSoc
      summary: Set battery buffer SoC
      description: "Set battery buffer SoC."
      externalDocs:
        url: https://docs.evcc.io/en/features/battery
      tags:
        - battery
      parameters:
        - $ref: "#/components/parameters/soc"
      responses:
        "200":
          $ref: "#/components/responses/NumberResult"
  /bufferstartsoc/{soc}:
    post:
      operationId: setBufferStartSoc
      summary: Set battery buffer start SoC
      description: "Set battery buffer start SoC."
      externalDocs:
        url: https://docs.evcc.io/en/features/battery
      tags:
        - battery
      parameters:
        - $ref: "#/components/parameters/soc"
      responses:
        "200":
          $ref: "#/components/responses/NumberResult"
...
  /prioritysoc/{soc}:
    post:
      operationId: setPrioritySoc
      summary: Set battery priority SoC
      description: "Set battery priority SoC."
      externalDocs:
        url: https://docs.evcc.io/en/features/battery
      tags:
        - battery
      parameters:
        - $ref: "#/components/parameters/soc"
      responses:
        "200":
          $ref: "#/components/responses/NumberResult"
  /residualpower/{power}:
    post:
      operationId: setResidualPower
      summary: Set grid residual power
      description: "Set grid connection operating point."
      externalDocs:
        url: https://docs.evcc.io/en/reference/configuration/site#residualpower
      tags:
        - battery
      parameters:
        - $ref: "#/components/parameters/power"
      responses:
        "200":
          $ref: "#/components/responses/NumberResult"
  /gridexportlimit/{power}:
    post:
      operationId: setGridExportLimit
      summary: Set grid export limit
      description: "Set the static grid export power limit in W used as optimizer constraint. 0 disables the limit. An active HEMS curtailment takes precedence."
      tags:
        - general
      parameters:
        - $ref: "#/components/parameters/power"
      responses:
        "200":
          $ref: "#/components/responses/NumberResult"
  /solaradjusted/{enable}:
    post:
      operationId: setSolarAdjusted
      summary: Adjust solar forecast
      description: "Adjust the solar forecast to real production data of the current day."
      tags:
        - experimental
      parameters:
        - $ref: "#/components/parameters/enable"
      responses:
        "200":
          $ref: "#/components/responses/BooleanResult"
...
  /session/{id}:
    parameters:
      - $ref: "#/components/parameters/id"
    put:
      operationId: updateSession
      summary: Update charging session
      description: "Update vehicle, loadpoint or odometer of a charging session. Only provided fields are changed; a null odometer clears the stored value."
      externalDocs:
        url: https://docs.evcc.io/en/features/sessions
      tags:
        - sessions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                vehicle:
                  $ref: "#/components/schemas/VehicleName"
                loadpoint:
                  $ref: "#/components/schemas/LoadpointName"
                odometer:
                  $ref: "#/components/schemas/Odometer"
      responses:
        "200":
          description: Success
    delete:
      operationId: deleteSession
      summary: Delete charging session
      description: "Delete charging session."
      externalDocs:
        url: https://docs.evcc.io/en/features/sessions
      tags:
        - sessions
      responses:
        "200":
          $ref: "#/components/responses/NullResult"
  /gridsessions:
    get:
      operationId: getGridSessions
      summary: Grid limitation sessions
      description: "Returns a list of HEMS grid limitation events."
      tags:
        - sessions
      parameters:
        - name: format
          in: query
          description: Response format (default json)
          schema:
            type: string
            enum:
              - csv
        - name: lang
          in: query
          description: Language (defaults to accept header)
          schema:
            type: string
            example: de
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GridSessions"
            text/csv:
              schema:
                description: Download csv-file
                type: string
                format: binary
...
components:
  schemas:
    ChangePassword:
      type: object
      properties:
        current:
          $ref: "#/components/schemas/Password"
        new:
          $ref: "#/components/schemas/Password"
...
    Password:
      description: Admin password
      type: string
...
  parameters:
...
    soc:
      name: soc
      description: SOC in %
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/Soc"
...
    costLimit:
      name: cost
      description: Cost limit in configured currency (default EUR) or CO2 limit in g/kWh
      example: 0.25
      in: path
      required: true
      schema:
        type: number
    enable:
      name: enable
      description: "Charging mode."
      in: path
      required: true
      schema:
        type: string
        enum:
          - "true"
          - "false"
          - "1"
          - "0"
  responses:
...
    NumberResult:
      description: Success - Number result
      content:
        application/json:
          schema:
            type: number
...
    BooleanResult:
      description: Success - Boolean result
      content:
        application/json:
          schema:
            type: boolean
...
    Unauthorized:
      description: Unauthorized - Login and try again
      content:
        text/plain:
          schema:
            type: string
            enum:
              - Unauthorized
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: auth
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        Long-lived API key (prefixed `evcc_`, managed via /auth/apikey).
        Grants access to /api/db/* without requiring the X-Admin-Password header.